+++ packages/chains/src/avalancheFuji.ts
@@ -11,6 +11,7 @@ export const avalancheFuji = {
   },
   rpcUrls: {
     default: { http: ['https://api.avax-test.network/ext/bc/C/rpc'] },
+    public: { http: ['https://api.avax-test.network/ext/bc/C/rpc'] },
   },
   blockExplorers: {
     etherscan: { name: 'SnowTrace', url: 'https://testnet.snowtrace.io' },