+++ packages/chains/src/hardhat.ts
@@ -11,5 +11,6 @@ export const hardhat = {
   },
   rpcUrls: {
     default: { http: ['http://127.0.0.1:8545'] },
+    public: { http: ['http://127.0.0.1:8545'] },
   },
 } as const satisfies Chain