+++ packages/chains/src/metisGoerli.ts
@@ -11,6 +11,7 @@ export const metisGoerli = {
   },
   rpcUrls: {
     default: { http: ['https://goerli.gateway.metisdevops.link'] },
+    public: { http: ['https://goerli.gateway.metisdevops.link'] },
   },
   blockExplorers: {
     default: {