+++ packages/chains/src/optimism.ts
@@ -17,6 +17,9 @@ export const optimism = {
     default: {
       http: ['https://mainnet.optimism.io'],
     },
+    public: {
+      http: ['https://mainnet.optimism.io'],
+    },
   },
   blockExplorers: {
     etherscan: {
@@ -24,8 +27,8 @@ export const optimism = {
       url: 'https://optimistic.etherscan.io',
     },
     default: {
-      name: 'Etherscan',
-      url: 'https://optimistic.etherscan.io',
+      name: 'Optimism Explorer',
+      url: 'https://explorer.optimism.io',
     },
   },
   contracts: {