+++ packages/connectors/src/injected.ts
@@ -237,11 +237,7 @@ export class InjectedConnector extends Connector<
                 chainId: id,
                 chainName: chain.name,
                 nativeCurrency: chain.nativeCurrency,
-                rpcUrls: [
-                  chain.rpcUrls.public?.http[0] ??
-                    chain.rpcUrls.default.http[0] ??
-                    '',
-                ],
+                rpcUrls: [chain.rpcUrls.public?.http[0] ?? ''],
                 blockExplorerUrls: this.getBlockExplorerUrls(chain),
               },
             ],