+++ packages/connectors/src/utils/getInjectedName.test.ts
@@ -31,6 +31,7 @@ describe.each([
   { ethereum: { isOneInchIOSWallet: true }, expected: '1inch Wallet' },
   { ethereum: { isOneInchAndroidWallet: true }, expected: '1inch Wallet' },
   { ethereum: { isPortal: true }, expected: 'Ripio Portal' },
+  { ethereum: { isRainbow: true }, expected: 'Rainbow' },
   { ethereum: { isTally: true }, expected: 'Tally' },
   {
     ethereum: { isTokenPocket: true, isMetaMask: true },