+++ packages/connectors/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @wagmi/connectors
 
+## 0.1.8
+
+### Patch Changes
+
+- 5e6dc30: Replaced legacy qrcodemodal with web3modal for WalletConnect v2.
+
+## 0.1.7
+
+### Patch Changes
+
+- be4add2: Added `isRainbow` flag to `InjectedConnector`.
+
+## 0.1.6
+
+### Patch Changes
+
+- 3dfc558: Add `switchSigner` method to `MockProvider`.
+
 ## 0.1.5
 
 ### Patch Changes